refactor: 모델 정의 위치 변경
