+++ src/client/shared/api/sequencify-blog/fetcher.ts
@@ -1,5 +1,5 @@
 import { generateFetcher } from "@/src/client/shared/lib/api-fetcher/generateFetcher";
-import { BlogContent } from "@/src/common/model/blog-parser.model";
+import { BlogContent } from "@/src/common/model/blog-content.model";
 
 const BASE_URL =
   typeof window !== "undefined" ? window.location.origin : process.env.API_URL;
