+++ src/common/model/blocks/index.ts
@@ -1,0 +1,1 @@
+export * from "./blog-block.model";
