+++ src/common/model/blocks/text-block.model.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const textBlockModel = z.object({
+  type: z.literal("text"),
+  value: z.string(),
+});
+
+export type TextBlock = z.infer<typeof textBlockModel>;
