+++ src/common/model/blog-content.model.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import { blogBlockModel } from "./blocks/blog-block.model";
+
+export const blogContentModel = z.object({
+  title: z.string(),
+  blocks: z.array(blogBlockModel),
+});
+
+export type BlogContent = z.infer<typeof blogContentModel>;
