+++ src/server/video-converter/video-converter.interface.ts
@@ -1,4 +1,4 @@
-import { BlogContent } from "@/src/common/model/blog-parser.model";
+import { BlogContent } from "@/src/common/model/blog-content.model";
 import { VideoRequestResult } from "@/src/common/model/video-request-result";
 
 export interface VideoConverter {
